Import and process AST data from an NCBI file

This function imports an antibiotic susceptibility testing (AST) dataset, processes the data, and optionally interprets the results based on MIC or disk diffusion data. It assumes that the input file is a delimited text file (e.g., CSV, TSV) and parses relevant columns (antibiotic names, species names, MIC or disk data) into suitable classes using the AMR package. It also uses the AMR package to determine susceptibility phenotype (SIR) based on EUCAST or CLSI guidelines

Usage

import_ncbi_ast(file, interpret = T)

Arguments

file

A string representing the path to the delimited file containing the AST data in NCBI antibiogram format. These files can be downloaded from the NCBI AST browser, e.g. https://www.ncbi.nlm.nih.gov/pathogens/ast#Pseudomonas%20aeruginosa

interpret

A logical value (default is TRUE). If TRUE, the function will interpret the susceptibility phenotype (SIR) for each row based on the MIC or disk diffusion values, and either EUCAST or CLSI testing standard (as indicated in the Testing standard column of the input file, if blank EUCAST will be used by default). If FALSE, no interpretation is performed.

Value

A data frame with the processed AST data, including additional columns:

biosample: The biological sample identifier (renamed from #BioSample).

spp_pheno: The species phenotype, formatted using the as.mo function.

drug_agent: The antibiotic used in the test, formatted using the as.ab function.

mic: The minimum inhibitory concentration (MIC) value, formatted using the as.mic function.

disk: The disk diffusion measurement (in mm), formatted using the as.disk function.

guideline: The guideline used for interpretation (either EUCAST or CLSI).

pheno: The interpreted phenotype (SIR) based on the MIC or disk diffusion data.

Examples

# Example usage
ast_data <- import_ncbi_ast("path/to/ast_data.tsv", interpret = TRUE)
#> Error in process_input(input): Input must be either a valid file path or a dataframe.
head(ast_data)
#> Error: object 'ast_data' not found